Fix optional and default args methods gen
