+++ buildSrc/src/main/kotlin/tool/generator/api/definition/node/transform/method/add methods for optional args.kt
@@ -3,7 +3,6 @@ package tool.generator.api.definition.node.transform.method
 import tool.generator.api.definition.node.DefinitionNode
 import tool.generator.api.definition.node.Nodes
 import tool.generator.api.definition.node.transform.TransformationChain
-import tool.generator.api.definition.node.type.method.ArgDefinitionNode
 import tool.generator.api.definition.node.type.method.ArgsDefinitionNode
 import tool.generator.api.definition.node.type.method.MethodDefinitionNode
 import tool.generator.api.definition.node.type.method.ext.*
@@ -32,7 +31,7 @@ object `add methods for optional args` : TransformationChain.Transform {
                 val args = node.signature.argsList
 
                 args.forEachIndexed { index, arg ->
-                    if (arg.hasFlag(ArgFlag.OPTIONAL) && isPossibleToCreateOptionalArgs(args, index, arg)) {
+                    if (arg.hasFlag(ArgFlag.OPTIONAL)) {
                         val newMethod = node.copy()
                         val newArgs = ArgsDefinitionNode()
 
@@ -50,28 +49,5 @@ object `add methods for optional args` : TransformationChain.Transform {
         }
 
         return result
-    }
-
-    /**
-     * It's impossible to create methods with optional arguments when neighboring arguments are of the same type.
-     */
-    private fun isPossibleToCreateOptionalArgs(
-        args: Collection<ArgDefinitionNode>,
-        index: Int,
-        arg: ArgDefinitionNode
-    ): Boolean {
-        if (index - 1 >= 0) {
-            val prevArg = args.toList()[index - 1]
-            if (prevArg.argType.type == arg.argType.type && prevArg.argType.flgas == arg.argType.flgas) {
-                return false
-            }
-        }
-        if (index + 1 < args.size) {
-            val nextArg = args.toList()[index + 1]
-            if (nextArg.argType.type == arg.argType.type && nextArg.argType.flgas == arg.argType.flgas) {
-                return false
-            }
-        }
-        return true
     }
 }
